fix(kit): unexport NewResults and Validate and fix test failure
This unexports/renames NewResults and Validate as newResults and
validate. This also reorganizes the tests into internal tests
and exported tests. This allows to still test internal logic and
remove useless validate() calls that caused a test failure due to
a negative score resulting from duplicate test scores. Negative
scores does not validate at the parsing stage, but we leverage
a negative score value to signal bad score entries. This is not
a problem in the deployed version of quickfeed since validation
is only done when parsing the score results from json. Once the
score object has been parsed and validates correctly there is no
longer any chance that an attacker is able to change it, other
than the quickfeed code itself.
