+++ .vscode/quickfeed-words.txt
@@ -1,6 +1,7 @@
 addstore
 agpath
 alecthomas
+Alloc
 allstudents
 allteachers
 apitest
